MapView: fix long-press/context-menu behavior
We had some issues with the useLongPress() hook, which would trigger a
'click' event at the end of a long-press, canceeling the intended
action. This functionality was removed from desktop, so the 'submit'
screen can only be brought up by right-click, or through the UI button.

There were also some phantom touches when panning or zooming the map on
mobile, which would erroneously trigger handleAddMarker(). This was
fixed by detecting the number of active touches; only single-touch
long-presses are accepted to trigger a long-press.