mapActions: add map initialization and geolocation actions
On App load, we now initialize the map by checking the geolocation
permissions and preferred use of geolocation. The truth of both of these
is inferred as 'shouldUseGeolocation', which tells the map to use the
user's current position by default.

We also set up a listener for changes in geolocation permissions, so
that if a user removes permission during a session, we are sure to
update behavior and UI by updating 'shouldUseGeolocation'.