Coordinate: a utility class for working with coordinate pairs
This utility can be used to provide a standard interface for working
with coordinates of several varieties, including '[lon, lat]', '{ lon,
lat }', '{ latitude, longitude }', and others. These can be easily cast
to the other types with the asArray() and asObject() methods.

It also exposes a static method to find the distance in feet between two
coordinates, any may be expanded with others later.